Fixed the sorting of chats, made the getter functions in ChatService more sensible
Pulled the chat comparator into a separate object so we can use it to sort in onRestart too, made a small change so that the chats without anything to read are below the others.
The changes in the service are such that we can - from the activity - set the person we're talking to once and then the other methods automatically apply to their chat object.