Added everything that was missing for persistent message storage:
Added serialization and parsing for whole message threads to the MessageParser and the appropriate writing procedures to the ChatHolder.

Redid the reading because there seems to be no proper way to find out how large the file is without using the File class, and thus no easy way of reading the whole file. FileInputStream.toString() prints the string representation of the FileInputStream object, not the actual contents of the file.

Added a lot of logging about Activity lifetimes because I had trouble figuring out why the service wouldn't shut down. Now it does and saves chats and even reads them back on startup. It also still creates Max Problem and Hans Muster, so it will look like it read it twice but actually it's just the old guys from the file and the new ones from onCreate()